+++ resources/views/components/share-recipient-list.blade.php
@@ -1,15 +1,16 @@
 @props([
     'recipients' => [],
-    'empty' => __('ui.no_brand_neighbors_are_available_92bb8ec323'),
+    'emptyTitle',
+    'empty',
 ])
 
-<div role="list" {{ $attributes->class(['share-recipient-list']) }}>
+<div role="list" data-share-recipients {{ $attributes->class(['share-recipient-list']) }}>
     @forelse ($recipients as $recipient)
         <x-share-recipient-item :recipient="$recipient" role="listitem" />
     @empty
         <x-empty-state
             icon="users"
-            title="{{ __('ui.no_neighbors_to_send_to_9a3ba8c390') }}"
+            :title="$emptyTitle"
             :description="$empty"
             compact
             role="listitem"